refactor(agents): separate copy rendering from sanitization
