+++ src/agents/embedded-agent-utils.ts
@@ -15,7 +15,7 @@ import {
   sanitizeAssistantVisibleText,
 } from "../shared/text/assistant-visible-text.js";
 import { sanitizeUserFacingText } from "./embedded-agent-helpers/sanitize-user-facing-text.js";
-import { renderUserFacingText } from "./failover/user-copy.js";
+import { renderUserFacingText } from "./embedded-agent-helpers/user-facing-text.js";
 import type { AgentMessage } from "./runtime/index.js";
 import { formatToolDetail, resolveToolDisplay } from "./tool-display.js";
